feat: require correct token to post and delete blogs. Tests are broke

+++ osa4/blogilista/controllers/blogs.js
@@ -1,10 +1,13 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
+
 blogsRouter.get('/', async (request, response) => {
+
   const blogs = await Blog
-  .find({}).populate('user', { username: 1, name: 1 })
+    .find({}).populate('user', { username: 1, name: 1 })
 
   if(blogs) {
     response.json(blogs)
@@ -15,15 +18,15 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const token = request.token
 
-  if(!body.userId){
-    throw {
-      name: 'ValidationError',
-      message: 'userId must be defined'
-    }
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+
+  if(!decodedToken || !decodedToken.id) {
+    response.status(401).json({ error: 'token missing or invalid' })
   }
 
-  const user = await User.findById(body.userId)
+  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     title: body.title,
@@ -35,7 +38,6 @@ blogsRouter.post('/', async (request, response) => {
 
   const result = await blog.save()
 
-  console.log(user)
   user.blogs = user.blogs.concat(result._id)
   await user.save()
 
@@ -44,9 +46,20 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   const idToDelete = request.params.id
-  console.log(idToDelete)
-  await Blog.findByIdAndRemove(idToDelete)
-  response.status(204).end()
+  const token = request.token
+  const user = request.user
+
+  const blog = await Blog.findById(idToDelete)
+
+  console.log(user)
+  if (user && blog && user._id.toString() === blog.user.toString()){
+    await Blog.findByIdAndRemove(idToDelete)
+    response.status(204).end()
+  } else if(!user) {
+    response.status(401).json({ error: 'Invalid token'})
+  } else if(!blog) {
+    response.status(404).json({ error: 'Invalid id'})
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
